refactor: move components to a components folder
